kfs: signals: Task termination and refactoring of code.
remove sig_eip and taskRegs from Task and Sigaction respectively.
Terminate child process on signal. Wrapper can be refactored to
reschedule, but used infinite while look as a POC. Change kill
syscall prototype to return i32 that gets stored in the tasks
eax and provide userspace POC that checks the return value of
the syscall and stops if it succeeded.

Signed-off-by: Vasileios Almpanis <vasilisalmpanis@gmail.com>