kfs: fs: Fix dentry caching problem.
When caching, parting of the hash is the name slice
This slice is passed as an argument by the user and
it resides on the stack. When we try to compare with
this slice the data on the stack can be already corrupted
from other calls. We should NEVER keep stack slices
saved in objects that outlive the function call that
passed the slice. Use the dentry name instead which is
a copy of that slice and is never modified. (For now).
Also override global inode number with the one comming
from the inode data for ext2. since it breaks caching.

Signed-off-by: Vasileios Almpanis <vasilisalmpanis@gmail.com>