kfs: mm: create memory mappings for userspace sections
instead of hardcoding the allocation of sections
and not tracking the memory anywhere use mmap_area
that automatically maps memory for use by userspace.
This is preliminary commit before we convert
virtual address space deletion to vma driven.
Currently we just iterate over all the PDE and PTE
to find what we need to remove.

Signed-off-by: Vasileios Almpanis <valmpanis@vasileios.almpanis>