kfs: bin: fix section allocation
when reading sections from the elf headers
we were using the absolute value of virt_address
that the start of the vma. our memory subsystem
expects addresses that are aligned to page size.
The bug was happening because we were setting a
non page aligned virtual address in our PD.

Signed-off-by: Vasileios Almpanis <vasilisalmpanis@gmail.com>